% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.r
\name{ripums_extra_example}
\alias{ripums_extra_example}
\title{Get path to ripums example datasets}
\usage{
ripums_extra_example(path = NULL)
}
\arguments{
\item{path}{Name of file. If `NULL`, the example files will be listed.}
}
\value{
The filepath to an example file, or if path is empty, a vector of all
  available files.
}
\description{
Get access to example extracts.
}
\examples{
ripums_extra_example() # Lists all available examples
ripums_extra_example("nhgis0010_csv.zip") # Gives filepath for a nhgis shape
}
