% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripumsexamples.r
\docType{package}
\name{ripumsexamples-package}
\alias{ripumsexamples}
\alias{ripumsexamples-package}
\title{Example data for the ripums package}
\description{
ripumsexamples is only used to bring a few data extrats from the ipums.org
website for use in examples and tests for the ripums package. These
files are too large to be included on CRAN and so have been put in
a separate package.
}
\author{
\strong{Maintainer}: Minnesota Population Center \email{gfellis@umn.edu}

}
